build vfd against 17.11 DPDK
Building VFd against 17.11 was failing. This patch fixes that.

+++ src/vfd/main.c
@@ -1291,7 +1291,7 @@ timeDelta(struct timeval * now, struct timeval * before)
 	we are currently managing.
 */
 void
-restore_vf_setings(uint8_t port_id, int vf_id) {
+restore_vf_setings(uint16_t port_id, int vf_id) {
 	int i;
 	int matched = 0;		// number matched for log
